move `Specification` to base module
